refactor(computer-use): neutralize plugin contract filename
