+++ turbo/apps/cli/src/lib/api/domains/zero-computer-use.ts
@@ -7,7 +7,7 @@ import type {
   ComputerUseReadCommandKind,
   ComputerUseWriteCommandKind,
 } from "@okouai/api-contracts/contracts/zero-computer-use";
-import type { ComputerUseAnyPluginCallBody } from "@okouai/api-contracts/contracts/zero-computer-use-plugins";
+import type { ComputerUseAnyPluginCallBody } from "@okouai/api-contracts/contracts/computer-use-plugins";
 import {
   zeroComputerUseAuthorizationRequestsContract,
   zeroComputerUseCommandContract,
