+++ turbo/apps/desktop/src/desktop-filesystem-plugin.ts
@@ -17,7 +17,7 @@ import {
   isComputerUseFilesystemTool,
   parseComputerUseFilesystemToolArguments,
   type ComputerUseFilesystemTool,
-} from "@okouai/api-contracts/contracts/zero-computer-use-plugins";
+} from "@okouai/api-contracts/contracts/computer-use-plugins";
 import type {
   ComputerUseCommand,
   ComputerUseCommandExecutionResult,
