+++ turbo/apps/desktop/src/desktop-mcp-plugin.ts
@@ -13,7 +13,7 @@ import {
   computerUseMcpServerCapability,
   isComputerUseMcpPluginCallPayload,
   type ComputerUseMcpPluginCallPayload,
-} from "@okouai/api-contracts/contracts/zero-computer-use-plugins";
+} from "@okouai/api-contracts/contracts/computer-use-plugins";
 import type {
   ComputerUseCommand,
   ComputerUseCommandExecutionResult,
